add custom Image component
Astro's Image component doesn't come with any blur placeholder. I tried to find
a library to help with this, but none of them worked. Instead, the following
gist showed how to recreate the Image component with a blur placeholder:

https://gist.github.com/sannajammeh/d1dfddc8c2538062de53714b07953cdf

If Astro ever adds a plur placeholder to their Image component, I will certainly
move over. I don't want to have to maintain this manually if I don't have to.